server/requirements.txt
torch
Pillow
transformers
fastapi
uvicorn
python-multipart